Q&A App using KMP's RESTFUL backend with review questions for the final test
Assignment Instructions:

Use the RESTful API described in class and documented at the link below to rewrite the Q&A application from A2 using AJAX. Feel free to reuse or adapt your code from A2 as you see fit if you want to but you don't have to. You also are free to recreate the DOM structure used in A2, but again you don't have to.

Your A4 version of the Q&A app should be accessible as a4.html in an a4 subdirectory in your class web space. Your app can be implemented using as many JavaScript and CSS resources as you need (including jQuery). Images are also fine to use. Your app, however, should only load once as a DOM, using client-side programming and AJAX to dynamically alter your interface and/or retrieve resources via the app's backend API as necessary.

Your app should include the following features:

Start with a login screen asking for username and password.
Provide feedback to the user if the login is incorrect and the ability to try again.
Upon successful login, show the user a "home" screen with all questions and answers.
Answered questions should be displayed with their answers.
There should be a way to delete the answer associated with a question, making the
question unanswered.
There should be a way to edit and update the answer associated with a question.
Unanswered questions should be displayed with some way to provide and submit an answer,
making the question answered.
There should be a way to go from the home screen to a "review" mode.
In review mode, a question along with a pair of answers to review should be retrieved from the backend.
A user should be able to register their relative grade between the two answers, and a new answer pair to review should be retrieved.
There should be a way to go back to the home screen from review mode.
Use CSS for styling and generally be laid out to fit on a screen (possibly with paginated or scrolling elements within the interface if needed).

Q&A app backend API documentation:

https://docs.google.com/document/d/1T3ZS2jRuTg5VTwzSBHws6a1RxENFTVia0kV-E2C3IZM/edit?usp=sharing